requirements.txt
# Web Scraping
playwright>=1.40.0
selenium>=4.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Proxy Management
fake-useragent>=1.4.0

# Database
psycopg2-binary
sqlalchemy>=2.0.0

# Data Processing
pandas
numpy
python-dateutil>=2.8.0

# NLP / Analysis (pure Python, no heavy builds)
vaderSentiment>=3.3.2

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
schedule>=1.2.0
retrying>=1.3.0
